Move Transaction related code to the transaction directory
